Fix EKF/UKF usage & move cd_dynamax utils
This streamlines use a bit and properly separates cd_dynamax utils from the rest of the library.
